chore: organize schema files
